### 2019-10-09 #1790 @whedon
[PRE REVIEW]: The ExoTETHyS package: Tools for Exoplanetary Transits around Host Stars
**Submitting author:** @gmorello (<a href="http://orcid.org/0000-0002-4262-5661">Giuseppe Morello</a>)
**Repository:** <a href="https://github.com/ucl-exoplanets/ExoTETHyS" target ="_blank">https://github.com/ucl-exoplanets/ExoTETHyS</a>
**Version:** v1.0.1
