### 2019-10-09 #1790 @arfon
:wave: @gmorello, thanks for your submission to JOSS. A few questions/comments before we get started here:

- Is this a joint submission with a AAS publication?
- Please look at this [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and format the YAML header _exactly_ the same way as it is done here. YAML is whitespace sensitive and your `paper.md` as it currently stands won't compile.
- Also, I don't see any tests/testing procedures included with this package. As such I don't think this is quite ready to send out to reviewers yet - if you would like to understand more about how your software will be reviewed, please take a look at our [reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).
