### 2019-10-09 25a4dac
Update sail_example2.txt
No output_path
