### 2019-10-09 8cda0d1
Update sail_example5.txt
No output_path
