### 2019-10-09 2f25fca
Update sail_example6.txt
No output_path
