### 2019-10-09 #1790 @gmorello
Dear Arfon Smith,

thanks for your message.
I edited the paper.md file to be identical to the example.

In case it does not compile:

  *   Could they be caused by accented letters?
  *   Not all the contributors have an ORCID, so I have reported only mine.
  *   Could they be caused by the URL that I report?

About the tests, there are already 8 possible input files in the examples folder.
Some of them require the user to change the paths in order to work on their computers, but at least the sail_example2.txt, sail_example5.txt, sail_example6.txt and sail_example7.txt should now work without any edits.

Please let me know if there are still issues.

Best regards,
Giuseppe
