+++ paper.md
@@ -32,7 +32,7 @@ affiliations:
  - name: Dept. Física Teórica y del Cosmos, Universidad de Granada, Campus de Fuentenueva s/n, 10871, Granada, Spain  
    index: 5  
  - name: Institut d'Astrophysique Spatiale, CNRS/Université Paris-Sud, Université Paris-Saclay, bâtiment 121, Université Paris-Sud, 91405, Orsay Cedex, France  
-   index: 6
+   index: 6  
  date: 8 October 2019  
 
  ---
