### 2019-10-09 33786f2
Fixing paper compilation
