+++ paper.md
@@ -1,45 +1,43 @@
 ---
-title: 'ExoTETHyS: Tools for Exoplanetary Transits around host stars'  
+title: 'ExoTETHyS: Tools for Exoplanetary Transits around host stars'
 tags:
- - python
- - astronomy
- - exoplanets
- - exoplanetary transits
- - stellar limb-darkening
-authors:  
- - name: Giuseppe Morello  
-   orcid: 0000-0002-4262-5661  
-   affiliation: "1, 2, 3"  
- - name: Antonio Claret    
-   affiliation: "4, 5"  
- - name: Marine Martin-Lagarde   
-   affiliation: 1  
- - name: Christophe Cossou   
-   affiliation: 6  
- - name: Angelos Tsiaras   
-   affiliation: 3  
- - name: Pierre-Olivier Lagage    
-   affiliation: 1  
-affiliations:  
- - name: AIM, CEA, CNRS, Université Paris-Saclay, Université Paris Diderot, Sorbonne Paris Cité, F-91191 Gif-sur-Yvette, France  
-   index: 1  
- - name: INAF, Osservatorio Astronomico di Palermo, Piazza del Parlamento 1, 90134 Palermo, Italy  
-   index: 2  
- - name: Dept. of Physics & Astronomy, University College London, Gower Street, WC1E 6BT London, UK  
-   index: 3  
- - name: Instituto de Astrofísica de Andalucía, CSIC, Apartado 3004, 18080 Granada, Spain  
-   index: 4  
- - name: Dept. Física Teórica y del Cosmos, Universidad de Granada, Campus de Fuentenueva s/n, 10871, Granada, Spain  
-   index: 5  
- - name: Institut d'Astrophysique Spatiale, CNRS/Université Paris-Sud, Université Paris-Saclay, bâtiment 121, Université Paris-Sud, 91405, Orsay Cedex, France  
-   index: 6  
- date: 8 October 2019  
+  - astronomy
+  - exoplanets
+  - exoplanetary transits
+  - galactic dynamics
+  - stellar limb-darkening
+authors:
+- name: Giuseppe Morello
+  orcid: 0000-0002-4262-5661
+  affiliation: "1, 2, 3"
+- name: Antonio Claret
+  affiliation: "4, 5"
+- name: Marine Martin-Lagarde
+  affiliation: 1  
+- name: Christophe Cossou
+  affiliation: 6  
+- name: Angelos Tsiaras
+  affiliation: 3  
+- name: Pierre-Olivier Lagage
+  affiliation: 1  
+affiliations:
+- name: "AIM, CEA, CNRS, Université Paris-Saclay, Université Paris Diderot, Sorbonne Paris Cité, F-91191 Gif-sur-Yvette, France"  
+  index: 1  
+- name: "INAF, Osservatorio Astronomico di Palermo, Piazza del Parlamento 1, 90134 Palermo, Italy"  
+  index: 2  
+- name: "Dept. of Physics & Astronomy, University College London, Gower Street, WC1E 6BT London, UK"  
+  index: 3  
+- name: "Instituto de Astrofísica de Andalucía, CSIC, Apartado 3004, 18080 Granada, Spain"
+  index: 4  
+- name: "Dept. Física Teórica y del Cosmos, Universidad de Granada, Campus de Fuentenueva s/n, 10871, Granada, Spain"  
+  index: 5  
+- name: "Institut d'Astrophysique Spatiale, CNRS/Université Paris-Sud, Université Paris-Saclay, bâtiment 121, Université Paris-Sud, 91405, Orsay Cedex, France"
+  index: 6
+date: 8 October 2019
+bibliography: paper.bib
+---
 
- ---
-
-#Summary 
- 
-Software URL: <https://github.com/ucl-exoplanets/ExoTETHyS>
+# Summary
 
[…]
