### 2019-10-09 b8ddcb5
Update paper.md
