### 2019-10-09 #1790 @arfon
@gmorello - this pull request should fix your paper: https://github.com/ucl-exoplanets/ExoTETHyS/pull/2

Also, please create a `paper.bib` file and add the bibtex entry for the arXiv entry in your paper. You will need to cite the arXiv entry in the text: you can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)).

> About the tests, there are already 8 possible input files in the examples folder.

OK thanks. This may be acceptable to our reviewers but I suspect they will want something more robust than this.

Next step is to identify some potential reviewers for this submission. Could you take a look at [this list of people](https://bit.ly/joss-reviewers) and identify a few people that might be able to give this package a review?
