### 2019-10-09 #1790 @arfon
> Is this a joint submission with a AAS publication?

Also, could you please answer this ☝️ question from earlier?
