### 2019-10-09 8b2cc50
Update paper.md
cite arxiv paper
