+++ paper.bib
@@ -1,0 +1,17 @@
+@ARTICLE{2019arXiv190809599M,
+       author = {{Morello}, Giuseppe and {Claret}, Antonio and {Martin-Lagarde}, Marine and
+         {Cossou}, Christophe and {Tsiaras}, Angelos and
+         {Lagage}, Pierre-Olivier},
+        title = "{The ExoTETHyS package: Tools for Exoplanetary Transits around Host Stars}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = "2019",
+        month = "Aug",
+          eid = {arXiv:1908.09599},
+        pages = {arXiv:1908.09599},
+archivePrefix = {arXiv},
+       eprint = {1908.09599},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190809599M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
