+++ paper.bib
@@ -1,4 +1,4 @@
-@ARTICLE{2019arXiv190809599M,
+@article{2019arXiv190809599M,
        author = {{Morello}, Giuseppe and {Claret}, Antonio and {Martin-Lagarde}, Marine and
          {Cossou}, Christophe and {Tsiaras}, Angelos and
          {Lagage}, Pierre-Olivier},
