### 2019-10-09 #1790 @gmorello
It is associated with an AAS manuscript under review, the one cited from arXiv.

Solely based on the information they provide, I selected possibly suitable reviewers:

nespinoza*, nicoguaro, andreas-h, zingale, ziotom78, steven-murray, williamjameshandley, harpolea, astrom-tom, danehkar, dhuppenkothen, wtbarnes, gregbryan

*nespinoza developed a code with similar functionalities. I am not sure whether this can cause a conflict of interest.

I have tried hard but I do not understand why the compiler does not recognise the citation @2019arXiv190809599M.

Thanks,
Giuseppe
