### 2019-10-18 6a3909a
Update trip.py
<> replaced with !=
