### 2019-10-23 #1834 @williamjameshandley
@arfon
> (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

I would consider myself an aspiring completist, but am currently not enough of one in order to disambiguate the above instruction. Could you point me to an example of how to link between a JOSS review and a GitHub issue?
