### 2019-10-29 ucl-exoplanets/ExoTETHyS#3 @steven-murray
Header of README is garbled
As part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834)

The header of the README seems to not be rendering correctly, especially the [//] notation.
