### 2019-10-29 ucl-exoplanets/ExoTETHyS#4 @steven-murray
dynamic MANIFEST
As part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834):

Creating a MANIFEST dynamically in a `setup.py` like in the following line:
https://github.com/ucl-exoplanets/ExoTETHyS/blob/6a3909a52dcea93ded6805d6220cc5247db4cae8/setup.py#L13
is to my knowledge highly uncommon practice. There are ways to dynamically include stuff in the MANIFEST file within the file specification itself.

Is there a good reason to do this?
