### 2019-10-29 ucl-exoplanets/ExoTETHyS#5 @steven-murray
Requirements
Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834):

The requirements as specified have no version specifications. It is typically helpful to have at least some minimum version specified:

https://github.com/ucl-exoplanets/ExoTETHyS/blob/6a3909a52dcea93ded6805d6220cc5247db4cae8/setup.py#L9
