### 2019-10-29 ucl-exoplanets/ExoTETHyS#6 @steven-murray
Installation instructions
Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834):

As far as I know, it is generally discouraged to install python packages with `python setup.py install` as this will not keep track of all installed files and so is not easily uninstallable. It is better to recommend installation using `pip install .`.

I will note that doing the following did *not* work for me:
```
conda create -n exotethys python=3
pip install . # from top-level directory of ExoTETHyS
```
but that might just be my own problem. Conda-installing `maptlotlib scipy astropy` then trying `pip install .` seems to have worked fine.

You may want to mention (depending on your intended audience) that creating a fresh environment (`venv` or `conda`) is encouraged.
