### 2019-10-29 ucl-exoplanets/ExoTETHyS#7 @steven-murray
Trying to run sail example
Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834):

I am trying to follow the examples on the README. The `sail` example gave me an error:

```
>>> sail.ldc_calculate('examples/sail_example1.txt`)
WARNING: /Users/pepe/Desktop/exotethys_versions/ExoTETHyS-master/Wavelength_bins_files/WFC3_G141_bins_Tsiaras2018_low.txt file not found.
WARNING: Ignoring wavelength_bins_file WFC3_G141_bins_Tsiaras2018_low.txt .
WARNING: assuming star_metallicity=0.0 for all targets. No input values were provided.
Downloading...  teff06000_logg4.0_MH0.0.pickle
Downloading...  teff06100_logg4.0_MH0.0.pickle
Downloading...  teff06000_logg4.5_MH0.0.pickle
Downloading...  teff06100_logg4.5_MH0.0.pickle
^[[OERROR: No legal passbands to calculate.
```

It seems that there is a hard-coded filepath in there?
