### 2019-10-29 ec6c6e8
Delete logo.png
