### 2019-10-29 ucl-exoplanets/ExoTETHyS#8 @steven-murray
Irregular imports
Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834):

This is a highly irregular/discouraged way to do imports:

https://github.com/ucl-exoplanets/ExoTETHyS/blob/ec6c6e8fb9d4fcc8f22c07556db2ce30c157c44b/exotethys/_1databases.py#L6

I recommend importing modules directly into your module.
