### 2019-10-29 ucl-exoplanets/ExoTETHyS#9 @steven-murray
Docstrings
Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834)

It seems that none of the functions in ExoTETHyS have docstrings, which makes it hard to know how to use it, beyond the examples in the README. Some of the functions do have comments which approximate a docstring, and at the very least, these should be converted to docstrings.
