### 2019-10-29 ucl-exoplanets/ExoTETHyS#3 @gmorello
Corrected, thanks
