### 2019-10-29 ucl-exoplanets/ExoTETHyS#10 @steven-murray
Tests
Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834):

There are no explicit tests for `ExoTETHys`. Automated tests (using eg. `pytest`) are very helpful in catching coding errors quickly, and also for helping new users know that their installation is working as intended.
