### 2019-10-29 ucl-exoplanets/ExoTETHyS#7 @gmorello
Yes, as many input files will be provided by the user it is normal to have hard-coded paths in the input files. However, the sail_example2.txt, sail_example5.txt, sail_example6.txt and sail_example7.txt should work without any edits.
