### 2019-10-29 ucl-exoplanets/ExoTETHyS#12 @steven-murray
Community Guidelines
Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834):

It is currently hard to know how to contribute to the repository. There are no explicit instructions, and the workings of the package are rather opaque (also due to #9).

It would be highly beneficial to create a simple `CONTRIBUTING.md` file, which outlines how to contribute to the software, how to make issues, etc.
