### 2019-10-29 ucl-exoplanets/ExoTETHyS#10 @gmorello
In principle, the examples are intended to be the tests. Some of them should work as they are, others require hard-coded paths, which is also something that should be tested. Do you have specific suggestions on how to improve these tests?
