### 2019-10-29 ucl-exoplanets/ExoTETHyS#13 @steven-murray
Code Description
Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834):

The current description of the project is this:

    This repository stores codes to help modeling exoplanetary transits and eclipsing binaries.

This could be both more explicit and more appealing, perhaps something like

    ExoTETHyS is an open-source package for modeling expolanetary transits and eclipsing binaries.    It provides pure-pythonic interfaces to the popular(?) `sail` and `trip` algorithms, along with a host of new improvements.
