### 2019-10-29 ucl-exoplanets/ExoTETHyS#3 @steven-murray
The lines

```
[//]: # (## Dependencies)

[//]: # (The code is consistent with python2/3. It makes use of)
[//]: # (- numpy, scipy, os, sys, glob, time, shutil, copy, pickle, astropy.io.fits, matplotlib)
```

are still missing from the rendering... is this intentional?
