+++ HISTORY.md
@@ -1,0 +1,10 @@
+## History of versions
+
+**version 1.0.0**  
+First online release
+
+**version 1.0.1**  
+* Added new database Atlas\_2000  
+* Phoenix\_2012\_13 and Phoenix\_2018 databases migrated to a different online directory  
+* Added check to passband wavelength limits  
+* Added error message if all required targets cannot be calculated  
