### 2019-10-29 ucl-exoplanets/ExoTETHyS#13 @gmorello
What about
`ExoTETHyS is an open-source package for modeling exoplanetary transits, eclipsing binaries and related phenomena.`

For example, just computing the limb-darkening coefficients is not "modeling a transit", but it is something needed to do that.
