### 2019-10-29 ucl-exoplanets/ExoTETHyS#7 @steven-murray
OK, I think I see. Perhaps it would be better to have an example on the README that worked out of the box without user input, or at least say that the file needs to be changed before running the example.
