### 2019-10-29 ucl-exoplanets/ExoTETHyS#10 @steven-murray
For your own sanity, it is useful to have this process automated. If you have an idea of what the output should be for these examples, you could literally just create a `tests` directory, write a `test_sail.py` module, and have a function per example. This can be set up to run automatically on every push at TravisCI.
