### 2019-10-29 ucl-exoplanets/ExoTETHyS#5 @gmorello
The code was originally written to work with quite old python2 libraries, but it may be quite annoying to dowgrade the libraries to see the minimum versions for the code to work. Could I just state the (new) versions for which it is guaranteed in the documentation, but without setting stringent requirements?
