### 2019-10-29 ucl-exoplanets/ExoTETHyS#3 @gmorello
Yes, it is. I was thinking about providing minimum versions as you state in #5. What about the solution that I propose in #5?
