### 2019-10-29 ucl-exoplanets/ExoTETHyS#5 @steven-murray
I don't think you have to find the absolute minimum required versions for the packages -- you can probably set them reasonably to whatever you have installed in your current environment. They can always be downgraded over time as well.
