### 2019-10-29 #1834 @steven-murray
Apart from the code issues, which I have opened in the repository itself, I have the following comments about the paper:

1. While there is a statement of what the software's purpose is, there is no additional context in the opening sentences (i.e. is this an important task for the field? What motivates the construction of this code?)
2. While there is a passing reference to the code being able to outperform other codes, there is no discussion of alternatives to this code, and what motivated the construction of this one, or whether this code is the _only_ way (currently) to use the SAIL and TRIP algorithms/packages.
3. Second sentence: "Now, " can be removed.
4. The last section could be moved to its own paragraph, concerning "how the code is currently being used"
5. It would be good to identify ways the code may be used outside of the current uses.
6. "more details can be found in..." -- perhaps it should be "Scientific details of the algorithms can be found in <>, and technical usage details can be found on GitHub"
