### 2019-10-30 f16c94f
Update sail.py
corrected mnor bug stellar_params_from_file_name
