### 2019-10-30 86ae727
Update trip.py
