### 2019-10-30 d37bbde
Update trip.py
corrected import and eccentricity != 0
