+++ exotethys/_1databases.py
@@ -23,14 +23,6 @@ else:
 import glob
 import time
 import shutil
-import astropy.io.fits as pyfits
-import numpy as np
-import os
-from scipy.optimize import minimize
-from scipy.interpolate import LinearNDInterpolator as interp
-from scipy.interpolate import interp1d
-
-import copy
 import pickle
 
 __all__ = ["databases"]
