### 2019-10-30 b9f9b71
Delete run_example.py
