+++ run_example.py
@@ -1,10 +1,0 @@
-from exotethys import sail, trip
-
-sail.ldc_calculate('tests/sail_example1.txt')
-#sail.ldc_calculate('tests/sail_example2.txt')
-#sail.ldc_calculate('tests/sail_example3.txt')
-#sail.ldc_calculate('tests/sail_example4.txt')
-#sail.ldc_calculate('tests/sail_example5.txt')
-#sail.ldc_calculate('tests/sail_example6.txt')
-#sail.ldc_calculate('tests/sail_example7.txt')
-#trip.trip_calculate('tests/trip_example.txt')
