### 2019-10-30 a70cfac
Update README.md
installing with pip
