### 2019-10-30 e1fef88
Update sail.py
trying docstring
