### 2019-10-30 ucl-exoplanets/ExoTETHyS#8 @gmorello
I checked all the import, reinstalled the code and it still works.
I think this issue can be closed.
