### 2019-10-30 625704d
Update sail.py
modified docstring
