### 2019-11-03 986f212
Update sail.py
Modified my_vstack such that the code will now crush if the input is not as expected.
