### 2019-11-04 ucl-exoplanets/ExoTETHyS#4 @steven-murray
Sorry for the late reply! Here's a pretty detailed list of suggestions for MANIFEST.in: https://www.reddit.com/r/Python/comments/40s8qw/simplify_your_manifestin_commands/
