### 2019-11-04 ucl-exoplanets/ExoTETHyS#8 @steven-murray
I think then that `_0imports.py` can be removed.
