### 2019-11-11 5d33802
Update sail.py
Docstrings up to the function check_2Darray
