### 2019-11-11 1a35ddd
Update sail.py
Docstrings up to the function get_grid_parameters
