### 2019-11-18 ucl-exoplanets/ExoTETHyS#7 @williamjameshandley
I also get a similar issue when trying to run the suggested example in the README:

```python
>>> from exotethys import trip
>>>trip.trip_calculate('examples/trip_example.txt')
WARNING: /Users/pepe/Desktop/aux_files/ld_Teff6100.0_logg4.5_MH0.0_TESS.txt file not found.
ERROR: invalid format for input_limb_file /Users/pepe/Desktop/aux_files/ld_Teff6100.0_logg4.5_MH0.0_TESS.txt . It must have 2 columns.
NameError                                 Traceback (most recent call last)
<ipython-input-2-3c574cbd2807> in <module>
----> 1 trip.trip_calculate('examples/trip_example.txt')

~/Projects/forks/ExoTETHyS/venv/lib/python3.8/site-packages/exotethys/trip.py in trip_calculate(configuration_file)
    952     [check, input_dict] = check_configuration(input_dict1) #Checking configuration file and dictionary update
    953     if check:
--> 954         process_configuration(input_dict) #Computing and saving
    955     else:
    956         print('Something went wrong with the input.')

~/Projects/forks/ExoTETHyS/venv/lib/python3.8/site-packages/exotethys/trip.py in process_configuration(input_dict)
    807     input_limb_path = input_dict_local['input_limb_path'][0]
    808     input_limb_file = input_dict_local['input_limb_file'][0]
--> 809     [mu_model_orig, radi_model_orig, intensity_model_orig] = get_input_limb_model(input_limb_type,input_limb_path,input_limb_file)
    810
    811     cutting_limb = input_dict_local['cutting_limb'][0]

~/Projects/forks/ExoTETHyS/venv/lib/python3.8/site-packages/exotethys/trip.py in get_input_limb_model(input_limb_type, input_limb_path, input_limb_file)
    711     if not check_2Darray(limb_model, n_col=2):
    712         print('ERROR: invalid format for input_limb_file', input_limb_path+input_limb_file, '. It must have 2 columns.')
--> 713         exit()
    714     intensity_model = copy.deepcopy(limb_model[:,1])
    715     if np.min(intensity_model)<0.0:

NameError: name 'exit' is not defined
```
