### 2019-11-18 ucl-exoplanets/ExoTETHyS#14 @williamjameshandley
Continuous integration
This project should have a rudimentary continuous integration set up which can be extended as time goes on.

In my experience, the simplest form of python CI is [Travis](https://docs.travis-ci.com/) combined with `pytest`. There are plenty of resources online, and I can point you to a few as necessary, but [this hello world example](https://github.com/kevchn/travis-ci-pytest) has served me pretty well in the past.

For now, a working travis setup which runs the two examples in the README would would suffice to resolve this issue. That setup would serve to (a) test the installation and (b) check that it runs without error.

Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834)
