### 2019-11-18 ucl-exoplanets/ExoTETHyS#15 @williamjameshandley
Make repository pip-installable
It would be good to twine upload this repository to [PyPi](https://pypi.org/) so that it is pip installable without having to clone the repository.

The tutorial on [python.org](https://packaging.python.org/tutorials/packaging-projects/) has served me well in the past for getting this up and running.

Part of JOSS [review](https://github.com/openjournals/joss-reviews/issues/1834)
