### 2019-11-18 #1834 @williamjameshandley
The repository installs fine on my machine (I'm using the recently released [python 3.8](https://docs.python.org/3/whatsnew/3.8.html), so that's a good sign given that I assume it hasn't necessarily been tested on this version).

- [ ]The [arXiv paper](https://arxiv.org/abs/1908.09599) makes it pretty clear that the intention is to extend this system of package as time goes on, which I approve of. My main requirement for acceptance is that a template for continuous integration is set up which can also be extended as the codebase expands. I've added this as a [GitHub issue](https://github.com/ucl-exoplanets/ExoTETHyS/issues/14), and we can discuss further there.

- [ ]It would also be good to upload this to [PyPi](https://pypi.org/) so that it can be installed without the need to clone the git repo. Also added as an [issue](https://github.com/ucl-exoplanets/ExoTETHyS/issues/15)

Other than this, more minor additions could be:

> 1. SAIL (Stellar Atmosphere Intensity Limb), i.e., a calculator of stellar limb-darkening coeffi-
cients that outperforms the existing software by one order of magnitude in terms of light-curve
model accuracy, i.e., down to <10 parts per million (ppm);
- [ ] This claim could be backed up by a specific example that the user can run

>2. TRIP (Transit Ring-Integrated Profile), which can compute an exact transit light-curve
by direct integration of the occulted stellar flux from the model intensities, without using a
parameterization (limb-darkening law) to approximate the stellar intensity profile.
- [ ] There should be a clear example of this in the README, ideally with a plot to show it working.
