### 2019-11-18 267ea27
Update sail.py
Docstrings up to get_limb_darkening_coefficients
