+++ exotethys/sail.py
@@ -936,8 +936,13 @@ def get_grid_parameters(stellar_models_grid):
 
 
 def get_subgrid(input_dict):
-#This function reads the file names from a database associated with a stellar_models_grid and extracts the stellar parameters.
-#Considers only the file names and parameters within the (optional) ranges specified in the configuration file.
+    """
+    This function extracts the file names, indices and stellar parameters from the selected database of stellar model-atmospheres within the selected (optional) parameter ranges.
+    
+    :param dict input_dict:
+    :return: the lists of database file names, the associated stellar parameters and file indices with the (optional) user parameter limits. It exits if the list of files is empty.
+    :rtype: list of str, np.array, np.array
+    """
     input_dict_local = copy.deepcopy(input_dict)
     input_keys = list(input_dict_local.keys())
     stellar_models_grid = input_dict_local['stellar_models_grid'][0]
@@ -1016,10 +1021,19 @@ def get_subgrid(input_dict):
 
 
 
-
-
 def get_neighbour_files_indices(target_name,star_effective_temperature,star_log_gravity,star_metallicity,star_params_grid,stellar_models_grid):
-#This function searchs for up to 2^n neighbours for a given set of stellar parameters and returns the corresponding indices in the files list.
+    """
+    This function searchs for up to 8 neighbours for a given set of stellar parameters and returns the corresponding indices in the files list (repeated indices are possible).
+    
+    :param str target_name: name associated with the target to compute
+    :param float star_effective_temperature: stellar effective temperature of the target
+    :param float star_log_gravity: stellar log gravity of the target
+    :param float star_metallicity: stellar metallicity of the target
+    :param np.array star_params_grid: 3-column array with stellar effective temperatures, log gravities and metallicities of the database models
+    :params str stellar_models_grid: the name of the chosen stellar database
+    :return: the indices of the (up to) eight nearest neighbour models (the array will contain 8 indices, but repeated indices are possible).
+    :rtype: np.array
+    """
     teff_min = np.min(star_params_grid[:,0])
     teff_max = np.max(star_params_grid[:,0])
     logg_min = np.min(star_params_grid[:,1])
@@ -1129,7 +1143,14 @@ def get_neighbour_files_indices(target_name,star_effective_temperature,star_log_
             
 
 def check_passband_limits(pb, stellar_models_grid):
-#This function checks that the wavelengths read from a passband file are within the limits for the chosen stellar_models_grid.
+    """
+    This function checks that the wavelengths read from a passband file are within the limits for the chosen stellar_models_grid.
+    
+    :param np.array pb: total spectral response table for the instrument/passband; the first column reports the wavelength in Angstrom, the second column is the electron/photon conversion factor or something proportional to this factor.
+    :params str stellar_models_grid: the name of the chosen stellar database
+    :return: the spectral response 2D array and a boolean value.
+    :rtype: np.array, bool
+    """
     check = True
     if stellar_models_grid == 'Phoenix_2018':
         minimum_wavelength = 500.0
@@ -1147,13 +1168,19 @@ def check_passband_limits(pb, stellar_models_grid):
         minimum_wavelength = 90.9
         maximum_wavelength = 1600000.0
         if np.min(pb[:,0])<minimum_wavelength or np.max(pb[:,0])>maximum_wavelength:
-            pb = False
+            check = False
         return pb, check
 
 
 def get_passband(passbands_path, passbands_ext, passband, stellar_models_grid):
-#This function extracts the passband from file; first column = wavelengths, second column = response
-#It returns the passband array and a boolean value
+    """
[…]
