### 2019-12-03 19d191c
Update sail.py
Docstrings completed
