+++ exotethys/tests/test_sail.py
@@ -2,6 +2,125 @@ import pytest
 from exotethys import sail
 import numpy as np
 
+
+testdata = [
+    ('ciao', 'ciao'),
+    ('5*5', '5*5'),
+    ('25', 25.0),
+    ('-3.14', -3.14),
+]
+
+@pytest.mark.parametrize('toread,reference', testdata)
+def test_str2float(toread, reference):
+    result = sail.str2float(toread)
+    np.testing.assert_equal(result, reference)
+
+
+#############################
+
+testdata = [
+    ('ciao', 'ciao'),
+    ('5*5', '5*5'),
+    ('25', 25.0),
+    ('-3.14', -3.14),
+]
+
+@pytest.mark.parametrize('toread,reference', testdata)
+def test_str2float(toread, reference):
+    result = sail.str2float(toread)
+    np.testing.assert_equal(result, reference)
+
+
+
+#############################
+
+testdata = [
+    (np.array([]), np.arange(5), np.arange(5)),
+    (np.arange(5), np.array([]), np.arange(5)),
+    (np.arange(5), np.arange(5)**2.0, np.array([[0, 1, 2, 3, 4], [0, 1, 4, 9, 16]])),
+    (np.ones((2,3)), np.arange(3), np.array([[1, 1, 1], [1, 1, 1], [0, 1, 2]])),
+]
+
+@pytest.mark.parametrize('array1,array2,reference', testdata)
+def test_my_vstack(array1, array2, reference):
+    result = sail.my_vstack(array1,array2)
+    np.testing.assert_equal(result, reference)
+
+def test_my_vstack_fails():
+    array1 = np.ones((2,3))
+    array2 = np.arange(2)
+    with pytest.raises(ValueError):
+        result = sail.my_vstack(array1,array2)
+
+
+#############################
+
+
+testdata = [
+    (np.array([]), {},  False),
+    (np.array([]), {'min_length': 0}, True),
+    (np.array([1]), {'max_length': 1}, True),
+    (np.array([1, 2]), {'max_length': 1}, False),
+]
+
+@pytest.mark.parametrize('vector,args,reference', testdata)
+def test_check_length(vector, args, reference):
+    result = sail.check_length(vector, **args)
+    np.testing.assert_equal(result, reference)
+
+
+
+#############################
+
+testdata = [
+    ([1, '2', 3], float, False),
+    ([1, '2', 3], str, False),
+    (['1', '2', '3'], str, True),
+    (['1', '2', '3'], float, False),
+    ([1., 2., 3.], float, True),
+]
+
+@pytest.mark.parametrize('vector,item_type,reference', testdata)
+def test_check_type_in_list(vector, item_type, reference):
+    result = sail.check_type_in_list(vector, item_type)
+    np.testing.assert_equal(result, reference)
+
+
+#############################
+
+testdata = [
+    (np.arange(6), {},  False),
+    (np.arange(6), {'min_value': 0}, True),
+    (np.arange(6)+1, {}, True),
+    (np.array([1, 2, 3.5]), {}, False),
+]
+
+@pytest.mark.parametrize('vector,args,reference', testdata)
+def test_check_integers(vector, args, reference):
+    result = sail.check_integers(vector, **args)
+    np.testing.assert_equal(result, reference)
+
+
+#############################
+
+testdata = [
+    (np.arange(6), {},  False),
+    (np.atleast_2d(np.arange(6)), {}, True),
+    (np.array([[1, 2, 3], [4, 5, 6]]), {}, True),
[…]
