### 2019-12-03 ucl-exoplanets/ExoTETHyS#10 @gmorello
I created the test_sail.py. Could you check if that is now fine?
(Meanwhile I will create the test_trip.py)
