### 2019-12-03 ucl-exoplanets/ExoTETHyS#9 @gmorello
Docstrings completed for sail.py.
I will soon start writing the docstrings for trip.py.
