### 2019-12-03 ucl-exoplanets/ExoTETHyS#8 @gmorello
Done, thanks
