### 2019-12-03 ucl-exoplanets/ExoTETHyS#10 @steven-murray
That file looks great! Are you going to try running a "full" example as a test as well? Do you need help setting up travis ci?
