### 2019-12-03 ucl-exoplanets/ExoTETHyS#9 @steven-murray
Those docstrings look good!
