### 2019-12-04 f28a8ea
Update sail.py
Removed * import
