### 2019-12-04 d4bcb2a
Update setup.py
added minimum versions to install_requires
