### 2019-12-04 ucl-exoplanets/ExoTETHyS#6 @gmorello
It works with `pip install .`
I do not often use environments, but I will consider it.
So maybe this issue can also be closed.
