### 2019-12-04 19fc453
Update setup.py
Updated install_requires
