+++ setup.py
@@ -6,7 +6,7 @@ import glob
 name = 'exotethys'
 description = ''
 url = ''
-install_requires = ['matplotlib>=2.0.0', 'docopt>=0.6.2', 'astropy>=2.0', 'scipy>=1.0.0', 'ephem>=3.7.7.0']
+install_requires = ['matplotlib>=2.2.4', 'scipy>=1.2.2, numpy>=1.16.5']
 
 os.chdir(os.path.abspath(os.path.dirname(__file__)))
 
