### 2019-12-04 5b088d7
Update setup.py
Updated install_requires
