### 2019-12-04 0593e4e
Update README.md
Added dependencies
