### 2019-12-04 ucl-exoplanets/ExoTETHyS#3 @gmorello
Finally, dependencies added.
This issue should be closed as well.
