### 2019-12-04 c30bf66
Update README.md
Correct dependencies
