### 2019-12-04 e06bcf5
Update setup.py
Correct dependencies
