### 2019-12-04 17f29a6
Update _database.py
comment matplotlib
