### 2019-12-04 5e4663d
Update sail.py
Comment matplotlib
