+++ exotethys/sail.py
@@ -4,12 +4,12 @@ from __future__ import print_function
 
 import os
 
-import matplotlib
-if os.environ.get('DISPLAY', '') == '':
-    print('no display found. Using non-interactive Agg backend')
-    matplotlib.use('Agg')
-else:
-    matplotlib.use('TkAgg')
+#import matplotlib
+#if os.environ.get('DISPLAY', '') == '':
+#    print('no display found. Using non-interactive Agg backend')
+#    matplotlib.use('Agg')
+#else:
+#    matplotlib.use('TkAgg')
 
 import pkg_resources
 import numpy as np
