### 2019-12-04 e135eee
Update trip.py
Comment matplotlib
