+++ .travis.yml
@@ -1,0 +1,9 @@
+language: python
+python:
+  - "2.7"
+  - "3.7"
+# command to install dependencies
+install:
+  - pip install .
+# command to run tests
+script: pytest
