+++ setup.py
@@ -1,66 +1,50 @@
-from setuptools import setup
-import codecs
+#!/usr/bin/env python
+
+"""
+Setup file
+"""
+
+import io
 import os
-import glob
+import re
 
-name = 'exotethys'
-description = ''
-url = ''
-install_requires = ['scipy>=1.2.2', 'numpy>=1.16.5']
+from setuptools import setup, find_packages
 
-os.chdir(os.path.abspath(os.path.dirname(__file__)))
 
-subdirs_to_include = []
-for x in os.walk(name):
-    if os.path.isdir(x[0]):
-        if x[0] != name:
-            subdirs_to_include.append(x[0])
+def read(*names, **kwargs):
+    with io.open(
+        os.path.join(os.path.dirname(__file__), *names),
+        encoding=kwargs.get("encoding", "utf8")
+    ) as fp:
+        return fp.read()
 
-files_to_include = []
-for x in glob.glob(os.path.join(name, '*')):
-    if os.path.isfile(x):
-        if x.split('.')[-1] not in ['py']:
-            files_to_include.append(os.path.join(name, os.path.split(x)[1]))
 
-files_to_include.append('README.md')
-files_to_include.append('LICENSE')
+def find_version(*file_paths):
+    version_file = read(*file_paths)
+    version_match = re.search(r"^__version__ = ['\"]([^'\"]*)['\"]",
+                              version_file, re.M)
+    if version_match:
+        return version_match.group(1)
+    raise RuntimeError("Unable to find version string.")
 
-w = open('MANIFEST.in', 'w')
-for i in subdirs_to_include:
-    w.write('include ' + os.path.join(i, '*') + ' \n')
-
-for i in files_to_include:
-    w.write('include ' + i + ' \n')
-
-w.close()
-
-with codecs.open('README.md', encoding='utf-8') as f:
-    long_description = f.read()
-
-version = ' '
-for i in open(os.path.join(name, '__init__.py')):
-    if len(i.split('__version__')) > 1:
-        version = i.split()[-1][1:-1]
 
 setup(
-    name=name,
-    version=version,
-    description=description,
-    long_description=long_description,
-    url=url,
+    name="exotethys",
+    version=find_version("exotethys", "__init__.py"),
+    description="",
+    url="https://github.com/ucl-exoplanets/ExoTETHyS",
     author='Giuseppe Morello',
     author_email='giuseppe.morello.11@ucl.ac.uk',
-    license='Creative Commons Attribution 4.0 International License',
+    license='GNU General Public License v3 (GPLv3)',
     classifiers=['Development Status :: 4 - Beta',
                  'Environment :: Console',
                  'Intended Audience :: Science/Research',
                  'Topic :: Scientific/Engineering :: Astronomy',
-                 'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
                  'Operating System :: MacOS :: MacOS X',
                  'Programming Language :: Python :: 3.7',
                  ],
-    packages=[name],
-    install_requires=install_requires,
-    include_package_data=True,
-    zip_safe=False,
+    packages=find_packages(),
+    package_data={"exotethys":["*.pickle", "Passbands/*.pass"]},
+    scripts=[],
+    data_files=[('', ['README.md'])]
 )
