+++ requirements.txt
@@ -1,0 +1,4 @@
+pytest
+setuptools
+scipy>=1.2.2
+numpy>=1.16.5
