### 2019-12-04 e01b782
Update .travis.yml
temporally comment requirements check
