### 2019-12-04 ucl-exoplanets/ExoTETHyS#10 @gmorello
Yes, for now I get the error "no environment variables set".
It appears that the error originates from importing libraries which are not installed by default, e.g. scipy.
