### 2019-12-04 ucl-exoplanets/ExoTETHyS#5 @steven-murray
This does not seem to be done to me: the `install_requires` keyword has been taken out of `setup.py`. This will break automatic installation using `pip`. You _can_ read the `requirements.txt` into your setup.py if you wish to reduce redundancy.
