### 2019-12-04 ucl-exoplanets/ExoTETHyS#6 @steven-murray
The README does now say to use `pip install .`, so I consider this good enough.
