### 2019-12-04 ucl-exoplanets/ExoTETHyS#14 @steven-murray
Looks like this is being addressed in #10
