### 2019-12-04 ucl-exoplanets/ExoTETHyS#10 @steven-murray
Can you point me to the travis log? You should probably add py3.6 and py3.7 to your test matrix as well.
