### 2019-12-05 ucl-exoplanets/ExoTETHyS#14 @gmorello
Yes, I actually followed the instructions from the example linked here before answering in #10.
Given the redundancy I close this issue so that we continue the discussion in #10.
