### 2019-12-05 97f3f14
Update setup.py
re-added install_requires
