### 2019-12-05 ucl-exoplanets/ExoTETHyS#5 @gmorello
I have just re-added the install_requires
