### 2019-12-05 ucl-exoplanets/ExoTETHyS#4 @gmorello
I modified the setup.py without creating the MANIFEST.in as it is unnecessary.
Is that fine for you?
