### 2019-12-05 ucl-exoplanets/ExoTETHyS#4 @steven-murray
absolutely.
