### 2019-12-05 ucl-exoplanets/ExoTETHyS#10 @steven-murray
Yes, thank you. It looks like the issue was related to #5 , which has now been fixed. However, the build is failing on pypy when it tries to build `scipy`. I've never used `pypi` before, so I'm not sure how much I can help on that. It looks like you might have to install `blas` via `apt` on travis.
