### 2019-12-05 #1834 @steven-murray
Currently the author is working on closing the issues that we have made in his repo -- most of them have now been addressed.
