### 2019-12-05 9e85953
Update .travis.yml
added python 3.6 and 3.7
