+++ .travis.yml
@@ -1,6 +1,8 @@
 language: python
 python:
   - "2.7"
+  - "3.6"
+  - "3.7"
   - "3.8"
   # PyPy versions
   - "pypy"   # currently Python 2.7.13, PyPy 7.1.1
