### 2019-12-05 8a17845
Update .travis.yml
before install blas lapack?
