+++ .travis.yml
@@ -7,6 +7,8 @@ python:
   # PyPy versions
   - "pypy"   # currently Python 2.7.13, PyPy 7.1.1
   - "pypy3"  # currently Python 3.6.1,  PyPy 7.1.1-beta0
+before_install:
+  - sudo apt-get install libblas-dev liblapack-dev
 # command to install dependencies
 install:
 #  - pip install -r requirements.txt
