### 2019-12-05 ucl-exoplanets/ExoTETHyS#10 @gmorello
Thank you very much. It passed the last test by installing `blas` and `lapack` in the travis file.
But I still need to add some tests for the trip subpackage.
