### 2019-12-07 e1a931c
Update trip.py
Docstring up to check_configuration
