### 2019-12-07 9362f14
Update trip.py
indentation line 145
