### 2019-12-10 92eee1d
Update trip.py
completed docstrings
