+++ exotethys/trip.py
@@ -710,43 +710,68 @@ def check_configuration(input_dict):
 
 
 def compute_z_sep(phi, inclination, sma_over_rs, eccentricity, arg_pericenter):
-#This function
-	theta = 2.0 * np.pi * phi
-	inclination = inclination * np.pi / 180.0
-	arg_pericenter = arg_pericenter * np.pi / 180.0
+    """
+    This function computes the sky-projected star-planet separation from the orbital phase vector and other parameters.
+    
+    :param np.array phi: 1D array with orbital phases, i.e., time in units of the orbital period, phase is integer at mid-transit
+    :param float inclination: orbital inclination in deg (inclination=90 if edge-on)
+    :param float sma_over_rs: orbital semimajor axis in units of the stellar radius
+    :param float eccentricity: orbital eccentricity
+    :param float arg_pericenter: argument of pericenter
+    :return: sky-projected star-planet separation at the given phase values
+    :rtype: np.array
+    """
+    theta = 2.0 * np.pi * phi
+    inclination = inclination * np.pi / 180.0
+    arg_pericenter = arg_pericenter * np.pi / 180.0
 
-	if eccentricity != 0:
-		n = len(theta)
-		E = np.zeros(n)
-		ecc2 = np.sqrt((1.0+eccentricity)/(1.0-eccentricity))
-		fref = np.pi/2.0 - arg_pericenter #setting reference point for the true anomaly
-		Eref = 2.0 * np.arctan(1.0/ecc2 * np.tan(fref/2.0))
-		if Eref < -np.pi/2.0:
-			Eref = Eref + 2.0*np.pi
-		Mref = Eref - (eccentricity * np.sin(Eref))
-		for i in range(n):
-			Mtmp = theta[i] +  Mref
-			Etmp = Mtmp
-			for j in range(10):
-				Etmp = Etmp + ((Mtmp + eccentricity*np.sin(Etmp) - Etmp) / (1.0-eccentricity*np.cos(Etmp)))
-			E[i] = Etmp
-		#calculating true anomaly
-		f = 2.0*np.arctan(ecc2*np.tan(E/2.0))
-		#calculating distance from true anomaly as fraction
-		r_frac = (1.0-eccentricity**2)/(1.0 + eccentricity*np.cos(f))
-		#computing z
-		z_sep = 1.0 - ((np.sin(inclination)**2.0) * (np.sin(f+arg_pericenter)**2))
-		z_sep = sma_over_rs*r_frac*np.sqrt(z_sep)
+    if eccentricity != 0:
+        n = len(theta)
+        E = np.zeros(n)
+        ecc2 = np.sqrt((1.0+eccentricity)/(1.0-eccentricity))
+        fref = np.pi/2.0 - arg_pericenter #setting reference point for the true anomaly
+        Eref = 2.0 * np.arctan(1.0/ecc2 * np.tan(fref/2.0))
+        if Eref < -np.pi/2.0:
+            Eref = Eref + 2.0*np.pi
+        Mref = Eref - (eccentricity * np.sin(Eref))
+        for i in range(n):
+            Mtmp = theta[i] +  Mref
+            Etmp = Mtmp
+            for j in range(10):
+                Etmp = Etmp + ((Mtmp + eccentricity*np.sin(Etmp) - Etmp) / (1.0-eccentricity*np.cos(Etmp)))
+            E[i] = Etmp
+        #calculating true anomaly
+        f = 2.0*np.arctan(ecc2*np.tan(E/2.0))
+        #calculating distance from true anomaly as fraction
[…]
