### 2019-12-10 8e903cd
Update sail.py
corrected docstring
