### 2019-12-10 42cf70f
Add files via upload
added tests for TRIP
