+++ exotethys/tests/test_trip.py
@@ -1,0 +1,192 @@
+import pytest
+from exotethys import trip
+import numpy as np
+
+
+testdata = [
+    ('ciao', 'ciao'),
+    ('5*5', '5*5'),
+    ('25', 25.0),
+    ('-3.14', -3.14),
+]
+
+@pytest.mark.parametrize('toread,reference', testdata)
+def test_str2float(toread, reference):
+    result = trip.str2float(toread)
+    np.testing.assert_equal(result, reference)
+
+
+#############################
+
+
+testdata = [
+    (np.array([]), {},  False),
+    (np.array([]), {'min_length': 0}, True),
+    (np.array([1]), {'max_length': 1}, True),
+    (np.array([1, 2]), {'max_length': 1}, False),
+]
+
+@pytest.mark.parametrize('vector,args,reference', testdata)
+def test_check_length(vector, args, reference):
+    result = trip.check_length(vector, **args)
+    np.testing.assert_equal(result, reference)
+
+
+
+#############################
+
+testdata = [
+    ([1, '2', 3], float, False),
+    ([1, '2', 3], str, False),
+    (['1', '2', '3'], str, True),
+    (['1', '2', '3'], float, False),
+    ([1., 2., 3.], float, True),
+]
+
+@pytest.mark.parametrize('vector,item_type,reference', testdata)
+def test_check_type_in_list(vector, item_type, reference):
+    result = trip.check_type_in_list(vector, item_type)
+    np.testing.assert_equal(result, reference)
+
+
+#############################
+
+testdata = [
+    (np.arange(6), {},  False),
+    (np.arange(6), {'min_value': 0}, True),
+    (np.arange(6)+1, {}, True),
+    (np.array([1, 2, 3.5]), {}, False),
+]
+
+@pytest.mark.parametrize('vector,args,reference', testdata)
+def test_check_integers(vector, args, reference):
+    result = trip.check_integers(vector, **args)
+    np.testing.assert_equal(result, reference)
+
+
+#############################
+
+testdata = [
+    (np.arange(6), {},  False),
+    (np.atleast_2d(np.arange(6)), {}, True),
+    (np.array([[1, 2, 3], [4, 5, 6]]), {}, True),
+    (np.array([[1, 2, 3], [4, 5, 6]]), {'n_col': 3}, True),
+    (np.array([[1, 2, 3], [4, 5, 6]]), {'n_col': 2}, False),
+]
+
+@pytest.mark.parametrize('vector,args,reference', testdata)
+def test_check_2Darray(vector, args, reference):
+    result = trip.check_2Darray(vector, **args)
+    np.testing.assert_equal(result, reference)
+
+
+#############################
+
+testdata = [
+    (np.arange(6),  True),
+    (np.atleast_2d(np.arange(6)), False),
+    (np.array([[1, 2, 3], [4, 5, 6]]), False),
+    (np.array([]), False),
+]
+
+@pytest.mark.parametrize('vector,reference', testdata)
+def test_check_1Darray(vector, reference):
+    result = trip.check_1Darray(vector)
+    np.testing.assert_equal(result, reference)
+
+
+#############################
+
+testdata = [
+    ('rp_over_rs',  None),
+    ('eccentricity', 0.0),
+]
+
[…]
