### 2019-12-10 64cfabe
Update trip.py
corrected indentation in compute_z_sep
