### 2019-12-10 ucl-exoplanets/ExoTETHyS#9 @gmorello
Docstrings are now completed also for trip.py.
Can this issue be closed?
