### 2019-12-10 ucl-exoplanets/ExoTETHyS#10 @gmorello
I think that tests are also complete now, and travis ci does not reveal any troubles.
@steven-murray Running a full example in the tests sounds tricky, as it would require to read files outside the package anyway. But I am going to clarify which examples should work and which ones need to be edited in the README.
I hope this issue can be closed as well.
