### 2019-12-10 d602280
Update sail.py
fixing sail_example6 using OrderedDict
