+++ examples/sail_example4.txt
@@ -21,9 +21,9 @@ passbands				Kepler
 #star_metallicity			0.00
 #
 ##for GRID calculation_type
-#star_minimum_effective_temperature	4000.0
-#star_maximum_effective_temperature	8000.0
-#star_minimum_log_gravity		0.0
-#star_maximum_log_gravity		6.0
+star_minimum_effective_temperature	5000.0
+star_maximum_effective_temperature	7000.0
+star_minimum_log_gravity		4.5
+star_maximum_log_gravity		4.5
 #star_minimum_metallicity		0.0
 #star_maximum_metallicity		0.0
