### 2019-12-10 fdbbff4
Update sail_example6.txt
