+++ examples/sail_example6.txt
@@ -1,5 +1,5 @@
 #Setting specific I/O paths		
-output_path				/Users/pepe/Desktop/exotethys_output/
+#output_path				/Users/pepe/Desktop/exotethys_output/
 ##GENERIC
 calculation_type			!grid		individual
 stellar_models_grid			!Phoenix_2018	Phoenix_2012_13		!Atlas_2000
@@ -8,7 +8,7 @@ limb_darkening_laws			claret4		!power2		!square_root	!quadratic	!linear		!gen_cl
 #gen_poly_orders				1 2 3 4 5 6
 #passbands_path				/Users/pepe/Desktop/exotethys_versions/ExoTETHyS-master/exotethys/Passbands/
 passbands				WFC3_G141	TESS
-wavelength_bins_path			/Users/pepe/Desktop/exotethys_versions/ExoTETHyS-master/Wavelength_bins_files/
+wavelength_bins_path			ExoTETHyS-master/Wavelength_bins_files/
 wavelength_bins_files			WFC3_G141_bins_Tsiaras2018_low.txt   no_bins
 user_output				basic		!complete
 #
