### 2019-12-10 a4933fa
Update sail_example7.txt
