### 2019-12-10 76f67ff
Update trip_example.txt
