### 2019-12-10 a342e65
Create CONTRIBUTING.md
