+++ CONTRIBUTING.md
@@ -1,0 +1,12 @@
+This open-source package is offered to the community as is.
+The authors are working to add new functionalities and to improve the existing ones. The GitHub repository contains the most updated version, while the last stable version is available on Zenodo (TO BE DONE).
+
+The users can contribute by:
+1. identifying bugs;
+2. proposing different options that better suit their specific needs;
+3. providing new information, such as stellar databases to be added;
+4. any other propositions.
+
+Currently there are two ways to communicate with the authors:
+1. opening issues in the GitHub repository;
+2. sending an e-mail to giuseppe.morello.11@ucl.ac.uk.
