### 2019-12-10 ucl-exoplanets/ExoTETHyS#12 @gmorello
I created a basic CONTRIBUTING.md file. Is that ok?
