### 2019-12-10 ucl-exoplanets/ExoTETHyS#15 @gmorello
I answered all the other issues. If they can be closed, I will consider this version as stable in order to upload to PyPi (and Zenodo, as recommended by the AAS reviewers).
