### 2019-12-11 ucl-exoplanets/ExoTETHyS#9 @steven-murray
Yup, great!
