### 2019-12-11 ucl-exoplanets/ExoTETHyS#10 @steven-murray
That sounds good. I'm happy for this to be closed, however I will say that it's quite fine to be reading files outside the package (so long as they're available to download in an automatic way, or if "mock" ones can be created in the test package). I'll leave that up to you. The note in the readme does sound important though!
