### 2019-12-11 ucl-exoplanets/ExoTETHyS#7 @gmorello
It looks my previous answer got lost: now all the examples should run without errors if following the instructions of the README.
