### 2019-12-11 ucl-exoplanets/ExoTETHyS#12 @gmorello
Ok, thanks. If you would like to point at some examples with other ways of contributing that I can consider, but at least this issue is not critical for releasing the first static version outside GitHub.
