### 2019-12-11 eceb0b9
Added travis badge to README.md
