### 2019-12-11 ucl-exoplanets/ExoTETHyS#7 @williamjameshandley
I currently get the error
```python
>>> from exotethys import trip
>>> trip.trip_calculate('examples/trip_example.txt')
WARNING: ExoTETHyS-master/aux_files/ld_Teff6100.0_logg4.5_MH0.0_TESS.txt file not found.
ERROR: invalid format for input_limb_file ExoTETHyS-master/aux_files/ld_Teff6100.0_logg4.5_MH0.0_TESS.txt . It must have 2 columns.
```
