### 2019-12-11 ucl-exoplanets/ExoTETHyS#7 @williamjameshandley
However, if I rename the repository as ExoTETHyS-master, and then run the command  from one directory above then it does run. Do you think it's possible to make the directories relative?

This command also seems to consume a lot of memory >10GB. Is this usual?
