### 2019-12-11 ucl-exoplanets/ExoTETHyS#7 @gmorello
TRIP does consume a lot of memory as by default it uses 10^5 interpolated points for calculating integrals. (We are working on some optimization, but it will take some time because the optimal sampling is case-dependent).
