### 2019-12-11 ucl-exoplanets/ExoTETHyS#7 @williamjameshandley
>About the error it depends on the name of the folder as @williamjameshandley noted. The way I download the repository it is automatically named ExoTETHyS-master, but apparently it is not always the case. This problem would not exist if I suggest running the code from the root directory.
I proposed running it from one level above to actually test the installation, otherwise python would import the local directory called exotethys. However, given that the user is not expected to modify the source code, this is probably an unnecessary precaution.
Would you prefer that I propose running the examples from the root directory?

I reckon it's probably more intuitive to run from within the directory, although I do acknowledge that this can cause import clashes.
