### 2019-12-11 cad2a59
Update sail_example5.txt
To be run from root
