+++ examples/sail_example5.txt
@@ -8,7 +8,7 @@ limb_darkening_laws			claret4		!power2		!square_root	!quadratic	!linear		!gen_cl
 #gen_poly_orders				1 2 3 4 5 6
 #passbands_path				/Users/pepe/Desktop/exotethys_versions/ExoTETHyS-master/exotethys/Passbands/
 passbands				WFC3_G141
-wavelength_bins_path			ExoTETHyS-master/Wavelength_bins_files/
+wavelength_bins_path			Wavelength_bins_files/
 wavelength_bins_files			WFC3_G141_bins_Tsiaras2018_low.txt
 user_output				basic		!complete
 #
