### 2019-12-11 196b4c9
Update sail_example6.txt
To be run from root
