### 2019-12-11 7dc89c6
Update trip_example.txt
To be run from root
