+++ examples/trip_example.txt
@@ -3,10 +3,10 @@ output_filename				trip_ld_Teff6100.0_logg4.5_MH0.0_TESS
 output_fileext				.pickle	.txt
 ##INPUT INFO
 input_limb_type				mu	!radi
-input_limb_path				ExoTETHyS-master/aux_files/
+input_limb_path				aux_files/
 input_limb_file				ld_Teff6100.0_logg4.5_MH0.0_TESS.txt
 input_series_type			!phi	time	!z_sep
-input_series_path			ExoTETHyS-master/aux_files/
+input_series_path			aux_files/
 input_series_file			time_series.txt
 rp_over_rs				0.12
 ##if not z_sep
