### 2019-12-11 ucl-exoplanets/ExoTETHyS#7 @gmorello
Ok, I modified the examples to be run from the root folder. Also changed the README and including the note about memory for the TRIP example.
Does it work now?
