### 2019-12-11 ucl-exoplanets/ExoTETHyS#7 @williamjameshandley
>Does it work now?

Yes : )

>Also changed the README and including the note about memory for the TRIP example.

This note should be directly below/above the second example in the [list of subpackages section](https://github.com/ucl-exoplanets/ExoTETHyS#list-of-subpackages)

Once that's done I'm happy for this issue to be closed.
