### 2019-12-11 ucl-exoplanets/ExoTETHyS#7 @steven-murray
I strongly feel that the examples should be runnable from wherever you are, as long as you pass in the path to the example file. It should not be dependent on being in any given folder.
