### 2019-12-11 ucl-exoplanets/ExoTETHyS#12 @steven-murray
Here's an example of what I consider to be a fairly complete contributing guide, that you'd probably have to prune a bit: https://github.com/RadioAstronomySoftwareGroup/pyuvdata/blob/master/.github/CONTRIBUTING.md
