### 2019-12-12 ucl-exoplanets/ExoTETHyS#7 @gmorello
> I strongly feel that the examples should be runnable from wherever you are, as long as you pass in the path to the example file. It should not be dependent on being in any given folder.

Yes, but the correct path must be written in the example file.
