### 2019-12-12 ucl-exoplanets/ExoTETHyS#7 @gmorello
Or maybe I did not fully get it.
The examples are simple text files, I do not want anything more complicated than that.
As they may contain the path to read other files, either the absolute path or the relative path from where you are running the code must be specified in the example file.
