### 2019-12-12 ucl-exoplanets/ExoTETHyS#7 @steven-murray
Three things:

1. YAML or TOML files are almost certainly going to serve you better than the .txt files you have -- there are well-used packages to read them robustly.
2. The paths in the config files should be able to be _either_ absolute or relative. If relative, they should be relative to the config file's location. In your examples, they should be relative, since you know the path relative to the file. This can easily be achieved using the `os.path` module.
3. You may want to consider installing the example data (not the example configs, but the data they use) along with the package, in which case the example data could be auto-loaded based on an import of the package. This is definitely optional, but worth considering.
