### 2019-12-12 ucl-exoplanets/ExoTETHyS#7 @gmorello
@steven-murray
Thanks for your new comments.
1. For now I prefer using file formats that are more known by common users, unless it is strictly necessary or there is a huge advantage in using other formats. I will look what these formats are though.
2. Similarly, for this point, when I download a package and try to execute it without changing settings, I like that files are saved in the local folder which contains the actual work, rather than in the root folder, which I prefer to keep as the original. So, the path can already be relative, but from the working directory. Can we just let this as it is for now? If I will find out that the majority of users prefer a different setting, I will modify it later.
3. I have seen some packages doing this trick, but with the current setup if the user follows the instructions should be able to get to the example results without any trouble. After that, users will want to specify their paths anyway. So, I do not see an advantage in this particular case.

Last comment: the AJ paper was accepted 1 week ago, and they are only waiting for the link to the static version on Zenodo to publish it. I do not feel it is the case to futher delay the publication for this issue.
