### 2019-12-13 ucl-exoplanets/ExoTETHyS#7 @steven-murray
@gmorello : you're right that it currently works for the example on the README. I was confused because of all the absolute paths in the `sail_example1.txt` file, but it seems that they're all commented out anyway. This could perhaps be cleaned up a bit -- users don't need to know your directory structure ;-).

I'm OK with this going forward and not being a hindrance to publication. However, I'd strongly suggest considering my suggestions further in the near future.
