### 2019-12-13 be30962
Update sail_example1.txt
cleaned
