### 2019-12-13 1aa4d4e
Update sail_example1.txt
cleaned
