### 2019-12-13 b6efb55
Update sail_example3.txt
cleaned
