+++ examples/sail_example3.txt
@@ -1,19 +1,19 @@
 #Setting specific I/O paths		
-#output_path				/Users/pepe/Desktop/exotethys_output/
+#output_path				/desired_absolute_or_relative_path/
 ##GENERIC
 calculation_type			!grid		individual
 stellar_models_grid			!Phoenix_2018	!Phoenix_2012_13		Atlas_2000
 limb_darkening_laws			claret4		!power2		!square_root	!quadratic	!linear		!gen_claret	!gen_poly
 #gen_claret_orders			1 2 3 4 5 6
 #gen_poly_orders				1 2 3 4 5 6
-#passbands_path				/Users/pepe/Desktop/exotethys_versions/ExoTETHyS-master/exotethys/Passbands/
+#passbands_path				/absolute_or_relative_path_to_user_passbands/
 passbands				TESS
-#wavelength_bins_path			/Users/pepe/Desktop/exotethys_versions/ExoTETHyS-master/Wavelength_bins_files/
+#wavelength_bins_path			/absolute_or_relative_path_to_user_wavelength_bins_files/
 #wavelength_bins_files			no_bins
 user_output				!basic		complete
 #
 ##for INDIVIDUAL calculation_type
-#targets_path				/Users/pepe/Desktop/target_list/
+#targets_path				/absolute_or_relative_path_to_target_list/
 #targets_file				target_list2a.txt
 #target_names				HD209458b
 star_effective_temperature		28300.0
