### 2019-12-13 c5fd301
Update sail_example4.txt
cleaned
