### 2019-12-13 b824f36
Update sail_example5.txt
cleaned
