### 2019-12-13 e21231f
Update sail_example6.txt
