### 2019-12-13 5f5e4ba
Update sail_example7.txt
cleaned
