### 2019-12-13 5b5ddfc
Update trip_example.txt
cleaned
