### 2019-12-13 ucl-exoplanets/ExoTETHyS#7 @gmorello
Examples cleaned, thanks. I am leaving the commented keywords anyway even if they are not used, just for convenience.
I will consider the most recent suggestions for version 2 (this is going to be 1.0.2).

Soon I will re-check one last time that all the examples work as expected, then proceed with uploading to Pypi and Zenodo.
