### 2019-12-15 339a03c
Update __init__.py
version 1.0.2
