+++ exotethys/__init__.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-__version__ = '1.0.1'
+__version__ = '1.0.2'
 
 from .sail import *
 from .trip import *
