+++ examples/sail_example8.txt
@@ -1,0 +1,29 @@
+#Setting specific I/O paths		
+#output_path				/desired_absolute_or_relative_path/
+##GENERIC
+calculation_type			!grid		individual
+stellar_models_grid			!Phoenix_2018	Phoenix_2012_13		!Atlas_2000
+limb_darkening_laws			claret4		!power2		!square_root	!quadratic	!linear		gen_claret	!gen_poly
+gen_claret_orders			1 2 3 4 5 6
+#gen_poly_orders				1 2 3 4 5 6
+#passbands_path				/absolute_or_relative_path_to_user_passbands/
+passbands				uniform_phoenix_2012_13
+wavelength_bins_path			Wavelength_bins_files/
+wavelength_bins_files			WFC3_G141_bins_Tsiaras2018_low.txt
+user_output				basic		!complete
+#
+##for INDIVIDUAL calculation_type
+targets_path				target_list/
+targets_file				target_list2a.txt
+#target_names				HD209458b	WASP43b
+#star_effective_temperature		6065.0	4400.0
+#star_log_gravity			4.36	4.65
+#star_metallicity			0.00
+#
+##for GRID calculation_type
+#star_minimum_effective_temperature	4000.0
+#star_maximum_effective_temperature	6000.0
+#star_minimum_log_gravity		4.5
+#star_maximum_log_gravity		4.5
+#star_minimum_metallicity
+#star_maximum_metallicity
