### 2019-12-15 ucl-exoplanets/ExoTETHyS#15 @gmorello
Added on Pypi, thanks.
One question: now that it is done, if I want to update just the README or the examples, do I have to:
1) repeat the whole upload procedure;
2) changing the version number?
