### 2019-12-15 ucl-exoplanets/ExoTETHyS#15 @williamjameshandley
Best practice is to re-upload to PyPi whenever you decide to update [the version number](https://semver.org/), which is typically at a slower rate than things like README updates or other small software alterations.
