+++ HISTORY.md
@@ -4,7 +4,16 @@
 First online release
 
 **version 1.0.1**  
+Version under revision by the [Journal of Open Source Software (JOSS)](https://joss.theoj.org/)
 * Added new database Atlas\_2000  
 * Phoenix\_2012\_13 and Phoenix\_2018 databases migrated to a different online directory  
 * Added check to passband wavelength limits  
 * Added error message if all required targets cannot be calculated  
+
+**version 1.0.2**  
+Version fully revised by the JOSS reviewers  
+* Added documentation with docstrings  
+* Added ``pytest`` tests and [Travis](https://docs.travis-ci.com/) continuous integration  
+* Fixed small issues and bugs  
+* Added history and community guidelines files  
+* Made repository pip-installable  
