### 2019-12-16 ucl-exoplanets/ExoTETHyS#15 @gmorello
Ok, thanks for the directions
