### 2020-01-06 #1834 @arfon
:wave: @gmorello - happy new year!

Could you give me an update on the status of your updates based on the reviewer feedback?
