+++ paper.bib
@@ -1,3 +1,21 @@
+@ARTICLE{kipping10,
+       author = {{Kipping}, David M. and {Tinetti}, Giovanna},
+        title = "{Nightside pollution of exoplanet transit depths}",
+      journal = {MNRAS},
+     keywords = {methods: analytical, techniques: photometric, occultations, planetary systems, infrared: general, Astrophysics - Earth and Planetary Astrophysics},
+         year = "2010",
+        month = "Oct",
+       volume = {407},
+       number = {4},
+        pages = {2589-2598},
+          doi = {10.1111/j.1365-2966.2010.17094.x},
+archivePrefix = {arXiv},
+       eprint = {0912.1133},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010MNRAS.407.2589K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @article{morello20,
        author = {{Morello}, Giuseppe and {Claret}, Antonio and {Martin-Lagarde}, Marine and
          {Cossou}, Christophe and {Tsiaras}, Angelos and
