+++ paper.md
@@ -39,11 +39,12 @@ bibliography: paper.bib
 
 # Summary
 
-ExoTETHyS is a python2/3 package which aims to provide a stand-alone set of tools for modeling spectro-photometric observations of the transiting exoplanets. Now, it contains the following subpackages:  
+ExoTETHyS is a python2/3 package which aims to provide a stand-alone set of tools for modeling spectro-photometric observations of the transiting exoplanets. The package is not a mere collection of pre-existing software, but will contain algorithms with new features as well as enhancements of existing codes. The first release contains the following subpackages:  
 1. SAIL (Stellar Atmosphere Intensity Limb), i.e., a calculator of stellar limb-darkening coefficients that outperforms the existing software by one order of magnitude in terms of light-curve model accuracy, i.e., down to <10 parts per million (ppm);  
-2. TRIP (Transit Ring-Integrated Profile), which can compute an exact transit light-curve by direct integration of the occulted stellar flux from the model intensities, without using a parameterization (limb-darkening law) to approximate the stellar intensity profile.
+2. TRIP (Transit Ring-Integrated Profile), which can compute an exact transit light-curve by direct integration of the occulted stellar flux from the model intensities, without using a parameterization (limb-darkening law) to approximate the stellar intensity profile. Following requests by users, we are adding a function to model eclipsing binaries.
+Scientific details of the algorithms can be found in a companion paper (@morello20, accepted by AJ), and technical usage details can be found on GitHub.
 
-More details are contained in a companion paper (@2019arXiv190809599M, under review) and in the technical documentation on GitHub. The package is currently being used with synthetic data in order to assess the overall uncertainty in exoplanet spectra obtained with the JWST and ARIEL instruments, including the uncertainties in the measured stellar parameters and in the stellar models. It is also being used for the analysis of real exoplanetary transits. Another project is the empirical validation of stellar limb-darkening models through the analysis of high-precision photometric observations obtained, in particular, with Kepler, K2, TESS, CHEOPS and PLATO.
+ExoTETHyS is currently being used with synthetic data in order to assess the overall uncertainty in exoplanet spectra obtained with the JWST and ARIEL instruments, including the uncertainties in the measured stellar parameters and in the stellar models (e.g., @sarkar19). It is also being used for the analysis of real exoplanetary transits (e.g., @tsiaras19). Another project is the empirical validation of stellar limb-darkening models through the analysis of high-precision photometric observations obtained, in particular, with Kepler, K2, TESS, CHEOPS and PLATO. We intend to include more stellar databases for this purpose. New tools are under development to estimate and correct for the exoplanet nightside pollution (significantly different treatment and formulas than discussed by @kipping10) and the stellar activity effects. Future work will include rotational oblateness and gravity darkening, tidal deformations, the possible presence of exomoons and exorings.
 
 
 # Acknowledgments  
