+++ paper.bib
@@ -15,7 +15,6 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2010MNRAS.407.2589K},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
 @article{morello20,
        author = {{Morello}, Giuseppe and {Claret}, Antonio and {Martin-Lagarde}, Marine and
          {Cossou}, Christophe and {Tsiaras}, Angelos and
@@ -33,7 +32,6 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190809599M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
 @ARTICLE{sarkar19,
        author = {{Sarkar}, Subhajit and {Madhusudhan}, Nikku and {Papageorgiou}, Andreas},
         title = "{JexoSim: a time-domain simulator of exoplanet transit spectroscopy with JWST}",
@@ -51,7 +49,6 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.491..378S},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
 @ARTICLE{tsiaras19,
        author = {{Tsiaras}, Angelos and {Waldmann}, Ingo P. and {Tinetti}, Giovanna and
          {Tennyson}, Jonathan and {Yurchenko}, Sergey N.},
