### 2020-01-09 #1834 @gmorello
@steven-murray<https://github.com/steven-murray>

I think I have addressed all your points except the first part of point 2:
> While there is a passing reference to the code being able to outperform other codes, there is no discussion of alternatives to this code

There are many limb-darkening calculators such as SAIL, the discussion of the performances is section 3.1 of the arXiv/AJ paper, but I would avoid to just report a long list of references with a negative statement without detailed explanations. I propose to refer to the AJ paper with the full discussion without additional comments, as it is now. Do you have another idea?

Best regards,
Giuseppe
