### 2020-01-09 #1834 @steven-murray
@gmorello is there an updated version of the paper I can read?

If you don't want to fully recapitulate the discussion concerning comparison to other codes in this paper, perhaps you could abbreviate it by saying something like "More details, and a comprehensive comparison to other relevant codes, can be found in ....".
