### 2020-01-10 60dfaa1
Update paper.bib
debugging
