+++ paper.bib
@@ -1,20 +1,3 @@
-@ARTICLE{kipping10,
-       author = {{Kipping}, David M. and {Tinetti}, Giovanna},
-        title = "{Nightside pollution of exoplanet transit depths}",
-      journal = {MNRAS},
-     keywords = {methods: analytical, techniques: photometric, occultations, planetary systems, infrared: general, Astrophysics - Earth and Planetary Astrophysics},
-         year = "2010",
-        month = "Oct",
-       volume = {407},
-       number = {4},
-        pages = {2589-2598},
-          doi = {10.1111/j.1365-2966.2010.17094.x},
-archivePrefix = {arXiv},
-       eprint = {0912.1133},
- primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2010MNRAS.407.2589K},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
 @article{morello20,
        author = {{Morello}, Giuseppe and {Claret}, Antonio and {Martin-Lagarde}, Marine and
          {Cossou}, Christophe and {Tsiaras}, Angelos and
@@ -30,39 +13,5 @@ archivePrefix = {arXiv},
        eprint = {1908.09599},
  primaryClass = {astro-ph.EP},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019arXiv190809599M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-@ARTICLE{sarkar19,
-       author = {{Sarkar}, Subhajit and {Madhusudhan}, Nikku and {Papageorgiou}, Andreas},
-        title = "{JexoSim: a time-domain simulator of exoplanet transit spectroscopy with JWST}",
-      journal = {MNRAS},
-     keywords = {space vehicles: instruments, infrared: planetary systems, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
-         year = "2020",
-        month = "Jan",
-       volume = {491},
-       number = {1},
-        pages = {378-397},
-          doi = {10.1093/mnras/stz2958},
-archivePrefix = {arXiv},
-       eprint = {1910.11344},
- primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.491..378S},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-@ARTICLE{tsiaras19,
-       author = {{Tsiaras}, Angelos and {Waldmann}, Ingo P. and {Tinetti}, Giovanna and
-         {Tennyson}, Jonathan and {Yurchenko}, Sergey N.},
-        title = "{Water vapour in the atmosphere of the habitable-zone eight-Earth-mass planet K2-18 b}",
-      journal = {Nature Astronomy},
-     keywords = {Astrophysics - Earth and Planetary Astrophysics},
-         year = "2019",
-        month = "Sep",
-       volume = {3},
-        pages = {1086-1091},
-          doi = {10.1038/s41550-019-0878-9},
-archivePrefix = {arXiv},
-       eprint = {1909.05218},
- primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019NatAs...3.1086T},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
