### 2020-01-10 bce5014
Update paper.md
debugging
