### 2020-01-10 6347c4b
Update paper.md
back to current final version
