### 2020-01-10 656efa5
Update paper.bib
back to current final version
