### 2020-01-29 #1834 @arfon
@gmorello - I've made a couple of minor updates to your paper for readability in https://github.com/ucl-exoplanets/ExoTETHyS/pull/18

Also, do you know that the status of the AAS companion paper is here? Before we can publish this we should add the AAS paper DOI to the YAML header of the `paper.md`.
