### 2020-01-29 #1834 @arfon
/ cc @crawfordsm for visibility on this submission that has now completed review and is accepted into JOSS.
