### 2020-02-11 #1834 @arfon
@gmorello - could you please merge this PR too? https://github.com/ucl-exoplanets/ExoTETHyS/pull/19 . It adds the AAS information to the JOSS paper.
