### 2020-02-11 #1834 @arfon
@gmorello - your paper is now published in JOSS :zap::rocket::boom:
